backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
google-generativeai==0.7.1
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
python-docx==1.1.0
pypdf2==3.0.1
